code/requirements.txt
numpy==1.23.2
opencv_python_headless==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.19.0
tensorboardX==2.6
torch==1.7.0
torchvision==0.8.0
tqdm==4.64.1
transformers==4.25.1
